+++ src/main/java/model/Model662.java
@@ -18,7 +18,7 @@ public class Model662 {
     private int id;
     private int orderNo;
     private String orderNumberIdentification;
-    private int referenceNo;
+    private long referenceNo;
     private String commodity;
     
     public int getId(){
@@ -45,11 +45,11 @@ public class Model662 {
         this.orderNumberIdentification = orderNumberIdentification;
     }
     
-    public int getReferenceNo(){
+    public long getReferenceNo(){
         return referenceNo;
     }
     
-    public void setReferenceNo(int referenceNo){
+    public void setReferenceNo(long referenceNo){
         this.referenceNo = referenceNo;
     }
